+++ src/modules/debugData/components/debugData/debugData.component.ts
@@ -1,7 +1,12 @@
-import {Component, ChangeDetectionStrategy, OnDestroy, OnInit, ChangeDetectorRef} from "@angular/core";
+import {Component, ChangeDetectionStrategy, OnDestroy, OnInit, ChangeDetectorRef, ElementRef} from "@angular/core";
 import {Subscription} from "rxjs";
 
 import {DebugDataEnabledService} from "../../services/debugDataEnabled/debugDataEnabled.service";
+
+/**
+ * Name of css class for enabled debug data
+ */
+ const ENABLED = 'enabled';
 
 /**
  * Component used for displaying debug data
@@ -26,33 +31,39 @@ export class DebugDataComponent implements OnDestroy, OnInit
 
     /**
      * Indication whether is debug data enabled
+     *
+     * @internal
      */
     public enabled: boolean = false;
 
     //######################### constructor #########################
     constructor(protected _debugDataEnabledSvc: DebugDataEnabledService,
-                protected _changeDetector: ChangeDetectorRef)
+                protected _changeDetector: ChangeDetectorRef,
+                protected _element: ElementRef<HTMLElement>)
     {
     }
 
     //######################### public methods - implementation of OnInit #########################
-    
+
     /**
      * Initialize component
      */
     public ngOnInit()
     {
         this.enabled = this._debugDataEnabledSvc.enabled;
+        this._setEnabledCssClass();
 
         this._debugDataEnabledChangeSubscription = this._debugDataEnabledSvc.enabledChange.subscribe(() =>
         {
             this.enabled = this._debugDataEnabledSvc.enabled;
+            this._setEnabledCssClass();
+
             this._changeDetector.detectChanges();
         });
     }
 
     //######################### public methods - implementation of OnDestroy #########################
-    
+
     /**
      * Called when component is destroyed
      */
@@ -61,4 +72,21 @@ export class DebugDataComponent implements OnDestroy, OnInit
         this._debugDataEnabledChangeSubscription?.unsubscribe();
         this._debugDataEnabledChangeSubscription = null;
     }
+
+    //######################### private methods #########################
+
+    /**
+     * Sets enabled css class according enabled state
+     */
+     private _setEnabledCssClass(): void
+     {
+         if(this.enabled)
+         {
+             this._element.nativeElement.classList.add(ENABLED);
+         }
+         else
+         {
+             this._element.nativeElement.classList.remove(ENABLED);
+         }
+     }
 }